services/matching-engine/requirements.txt
kafka-python
redis
fastapi
uvicorn[standard]
prometheus-client
